requirements.txt
paramiko~=2.7.2
PyQt5~=5.15.4
python-dotenv
